Added review list dialog
